Refatorado o sistema para fazer validações de dados e não dar exceção para o usuário, mesmo não passando dados.
Took 1 hour 10 minutes

+++ application/dataprovider/src/main/resources/db/migration/V1.0__create.sql
@@ -46,7 +46,7 @@ CREATE TABLE venda_itens_tbl
     id_discos      int,
     valor          decimal(12, 2),
     quantidade     int,
-    cash_back      int,
+    cash_back      decimal(12,2),
     PRIMARY KEY (id_venda_itens)
 ) ENGINE = InnoDB
   DEFAULT CHARSET = utf8;
